refactored functionality to map member cards into component - ViewMembers
