+++ pages/team.js
@@ -1,10 +1,11 @@
 import React from 'react';
 import Head from 'next/head';
+import Home from '.';
 
 export default function Team() {
   return (
     <><Head><title>Your Film Production Team</title></Head>
-      <h1>Team</h1>
+      <div><Home /></div>
     </>
   );
 }
